// src/app/shared/Model/BusinessImpactAnalizeUpdate.ts
export class BusinessImpactAnalizeUpdate
{
  biA_ID :Number;
  biA_NAME: string;
  biA_Process_Evaluated: string;
  biA_DepartmentId: number;
  biA_Owner: number;
  biA_FocalPoint: number;
  biA_Status: number;
  created_By: number;
  created_Date: string;
  modified_By: number;
  modified_Date: string;
  
}